+++ app/views/static_pages/top.html.slim
@@ -1,12 +1,16 @@
+.flex-grow
+  h1 みんなの家計簿
+  p みんなの家計簿は家計簿公開サービスです！
+  p 〜サービス説明文〜
+
+  = link_to 'はじめる（ユーザー登録)', new_user_path
+  = link_to 'ログイン', login_path
+  = link_to 'お試しログイン', '#'
+
+#hello
+  | {{message}}
+  app
+
+
 = javascript_pack_tag 'hello_vue'
-
-
-/ .flex-grow
-/   h1 みんなの家計簿
-/   p みんなの家計簿は家計簿公開サービスです！
-/   p 〜サービス説明文〜
-
-/   = link_to 'はじめる（ユーザー登録)', new_user_path
-/   = link_to 'ログイン', login_path
-/   = link_to 'お試しログイン', '#'
